allow callback to provide additional capabilities
in my case it'll be SIZE, but who knows what other people might want to
have added to the capabilities - what sense in providing extensibility
when you can't use the protocols inherent extenstion facility.

On a second note, it seems wrong to clain support of 8BITMIME. 8BITMIME
not only implies support for accepting 8 bit messages, it also implies
support for recoding them if they arrive using 8 bits but the next hop
doesn't support 8 bits.

On the other hand, MTAs really shouldn't touch the mails contents and
neither should we force other MTAs to do so, so sending 8BITMIME feels
kind of ok.

Exim does the same thing and if Phil says it's ok, then it probably is.