correctly extract arguments
the verbs in SMTP are case insenstive. The previous way of extracting
broke if the command given by the server wasn't in all uppercase.

Because the checks for the command are already done in a case
insensitive manner, the server understood the command, but then failed
to correctly extract the argument returning the whole line including the
lowercased command

(yeah. I can get away not properly escaping the command as
extractArguments is only called with trusted data that does not contain
special characters)