added totally asynchronous interface to data
the old data() method wanted all the data to send as a parameter. IMHO
it's more advantageous to do the sending of potentially huge amounts of
data in a more chunk-based fashion.

Now the old functionality is still available, but we also provide

beginData() which just sends DATA and nothing more
sendData() which can be called as often as needed
endData() which sends the final .

The old data() methods is now using the new asynchorous facility too to
prove that the change really works.

While implementing this, I noticed that the client is a bit brittle as
it doesn't do any socket draining handling at the moment (now it does
for data, but still). We'll have to look into this at a later time